Changed PositionComponent x and y from float to double.

+++ include/ECS/Component/PositionComponent.h
@@ -5,11 +5,11 @@ namespace ECS::Component
     class PositionComponent : public Component
     {
     public:
-            PositionComponent(float x, float y) : Component("Position"), x(x), y(y) {}
+            PositionComponent(double x, double y) : Component("Position"), x(x), y(y) {}
             ~PositionComponent() {}
 
-            float x;
-            float y;
+            double x;
+            double y;
     };
 }
